initial(forms): create SessionForm component without terms :tada:
