+++ components/Form/index.ts
@@ -1,2 +1,3 @@
 export { default as ClassForm } from "./ClassForm";
+export { default as SessionForm } from "./SessionForm";
 export { default as SubjectForm } from "./SubjectForm";
